# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger IAA - OpenAPI 3.0
  version: 0.0.1
servers:
  - url: http://localhost:8080/rest
...
paths:
...
  /publication/{key}:
    delete:
      tags:
        - Publications
      summary: Delete a publication
      operationId: deletePublication
      parameters:
        - name: key
          in: path
          description: Key of the publication
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
...
  /assignment/{publicationKey}:
    get:
      tags:
        - Assignment
      summary: Get all existing assignment
      description: Get all existing assignment
      operationId: getAssignmentByPubKey
      parameters:
        - name: publicationKey
          in: path
          description: Key of the publication
          required: true
          schema:
            type: string
        - name: showReturned
          in: query
          description: show returned assignments
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successful assignment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
  /assignment/return/{uuid}:
    post:
      tags:
        - Assignment
      summary: return the assignment
      operationId: returnAssignment
      parameters:
        - name: uuid
          in: path
          description: Uuid of the assignment
          required: true
          schema:
            type: string
        - name: dateOfReturn
          in: query
          description: returndate
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful assignment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
  /assignment/extend/{uuid}:
    post:
      tags:
        - Assignment
      summary: extend the assignment
      operationId: extendAssignment
      parameters:
        - name: uuid
          in: path
          description: Uuid of the assignment
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful assignment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
  /assignment/publication-lost/{uuid}:
    post:
      tags:
        - Assignment
      summary: mark the assignment as lost
      operationId: publicationLostAssignment
      parameters:
        - name: uuid
          in: path
          description: Uuid of the assignment
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful assignment
  /assignment/latest-return-date:
    get:
      tags:
        - Assignment
      summary: returns the return date for a given date or now
      operationId: getAssignmentReturnDate
      parameters:
        - name: dateOfAssignment
          in: query
          description: Date of assignmnet
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  dateOfAssignment:
                    type: string
                    format: date-time
                  latestReturnDate:
                    type: string
                    format: date-time
...
  /kind-of-publication/{uuid}:
    delete:
      tags:
        - Kind of publications
      summary: Delete a kind of publication
      operationId: deleteKindOfPublication
      parameters:
        - name: uuid
          in: path
          description: Uuid of the kind
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
...
components:
  schemas:
    KindOfPublication:
      type: object
      properties:
        uuid:
          type: string
        value:
          type: string
    Author:
      type: object
      properties:
        uuid:
          type: string
        surname:
          type: string
        name:
          type: string
    Keyword:
      type: object
      properties:
        uuid:
          type: string
        value:
          type: string
...
    Borrower:
      type: object
      properties:
        uuid:
          type: string
        surname:
          type: string
        name:
          type: string
        studentNumber:
          type: string
...
    Publication:
      type: object
      required:
        - key
        - titel
      properties:
        key:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        dateOfPublication:
          type: string
          format: date-time
        publisher:
          type: string
        kindOfPublication:
          $ref: '#/components/schemas/KindOfPublication'
        isbn:
          type: string
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/Keyword'
        quantity:
          type: integer
          format: uint
          default: 0
          example: 69
    Assignment:
      type: object
      properties:
        uuid:
          type: string
        dateOfAssignment:
          type: string
          format: date-time
        dateOfReturn:
          type: string
          format: date-time
        extension:
          type: integer
          format: uint
          default: 0
          example: 2
        borrowedUtil:
          type: string
          format: date-time
        publication:
          $ref: '#/components/schemas/Publication'
        borrower:
          $ref: '#/components/schemas/Borrower'
...